canvas2d/fill-and-stroke-styles/2d.gradient.object.invalidoffset.html

<!DOCTYPE html>
<title>Canvas test: 2d.gradient.object.invalidoffset</title>
<meta name="author" content="Philip Taylor">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../../common/canvas-tests.js"></script>
<link rel="stylesheet" href="../../../common/canvas-tests.css">
<body class="show_output">

<h1><a href="index.2d.html">2d</a>.<a href="index.2d.gradient.html">gradient</a>.<a href="index.2d.gradient.object.html">object</a>.invalidoffset</h1>
<p class="desc"></p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.2d.gradient.invalidoffset">2d.gradient.invalidoffset</a>
</ul>
</div>

<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("");
_addTest(function(canvas, ctx) {

var g = ctx.createLinearGradient(0, 0, 100, 0);
assert_throws("INDEX_SIZE_ERR", function() { g.addColorStop(-1, '#000'); });
assert_throws("INDEX_SIZE_ERR", function() { g.addColorStop(2, '#000'); });
assert_throws("INDEX_SIZE_ERR", function() { g.addColorStop(Infinity, '#000'); });
assert_throws("INDEX_SIZE_ERR", function() { g.addColorStop(-Infinity, '#000'); });
assert_throws("INDEX_SIZE_ERR", function() { g.addColorStop(NaN, '#000'); });


});
</script>
